Moved merge config function into separate service
